jwt token added to all the controllers of maintenance service.

+++ MaintenanceWebApi/Controllers/EvaluationsController.cs
@@ -6,6 +6,9 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MaintenanceWebApi.Models;
+using JwtAuthenticationManager.Models;
+using JwtAuthenticationManager;
+using System.Net.Http;
 
 namespace MaintenanceWebApi.Controllers
 {
@@ -14,62 +17,111 @@ namespace MaintenanceWebApi.Controllers
     public class EvaluationsController : ControllerBase
     {
         private readonly MaintenanceDbContext _context;
+        private readonly JwtTokenHandler _JwtTokenHandler;
+        private readonly HttpClient _httpClient;
+        private readonly ILogger<EvaluationsController> _logger;
 
-        public EvaluationsController(MaintenanceDbContext context)
+        public EvaluationsController(MaintenanceDbContext context,HttpClient httpClient, ILogger<EvaluationsController> logger)
         {
             _context = context;
+            _httpClient = httpClient;
+            _logger = logger;
         }
 
         // GET: api/Evaluations
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Evaluation>>> Getevaluations(string companyId)
+        public async Task<ActionResult<IEnumerable<Evaluation>>> Getevaluations()
         {
-            return await _context.evaluations.Where(x=>x.companyId==companyId).ToListAsync();
+            try
+            {
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+
+                    return await _context.evaluations.Where(x => x.companyId == claimresponse.companyId).ToListAsync();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // GET: api/Evaluations/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Evaluation>> GetEvaluation(int id,string companyId)
+        public async Task<ActionResult<Evaluation>> GetEvaluation(int id)
         {
-            var evaluation = await _context.evaluations.Where(x=>x.evaluationAutoId==id && x.companyId==companyId).ToListAsync();
-
-            if (evaluation == null)
+            try
             {
-                return NotFound();
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+
+                    var evaluation = await _context.evaluations.Where(x => x.evaluationAutoId == id && x.companyId == claimresponse.companyId).ToListAsync();
+
+                    if (evaluation == null)
+                    {
+                        return NotFound();
+                    }
+                    return Ok(evaluation);
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
-            return Ok(evaluation);
         }
 
         // PUT: api/Evaluations/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut]
-        public async Task<IActionResult> PutEvaluation(int id,string companyId, Evaluation evaluation)
+        public async Task<IActionResult> PutEvaluation(int id,Evaluation evaluation)
         {
-            if (evaluation.evaluationAutoId==id && evaluation.companyId==companyId)
-            {
-
-                _context.Entry(evaluation).State = EntityState.Modified;
-
-            }
-
             try
             {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!EvaluationExists(id))
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+                    if (evaluation.evaluationAutoId == id && evaluation.companyId == claimresponse.companyId)
+                    {
 
-            return NoContent();
+                        _context.Entry(evaluation).State = EntityState.Modified;
+
+                    }
+
+                    try
+                    {
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!EvaluationExists(id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+
+                    return NoContent();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // POST: api/Evaluations
@@ -77,50 +129,64 @@ namespace MaintenanceWebApi.Controllers
         [HttpPost]
         public async Task<ActionResult<Evaluation>> PostEvaluation(Evaluation evaluation)
         {
-            var compId = _context.evaluations.Where(eval => eval.companyId == evaluation.companyId && eval.woAutoId == evaluation.woAutoId).Select(d => d.evaluationId).ToList();
-            var autoId = "";
-            if (compId.Count > 0)
+            try
             {
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    var compId = _context.evaluations.Where(eval => eval.companyId == claimresponse.companyId && eval.woAutoId == evaluation.woAutoId).Select(d => d.evaluationId).ToList();
+                    var autoId = "";
+                    if (compId.Count > 0)
+                    {
 
-                autoId = compId.Max(x => int.Parse(x.Substring(4))).ToString();
+                        autoId = compId.Max(x => int.Parse(x.Substring(4))).ToString();
+                    }
+
+                    if (autoId == "")
+                    {
+                        _context.ChangeTracker.Clear();
+                        Evaluation eval = new Evaluation();
+                        string comId = "EVAL1";
+                        eval.evaluationAutoId = evaluation.evaluationAutoId;
+                        eval.evaluationId = comId;
+                        eval.woAutoId = evaluation.woAutoId;
+                        eval.userName = evaluation.userName;
+                        eval.topName = evaluation.topName;
+                        eval.startTime = evaluation.startTime;
+                        eval.endTime = evaluation.endTime;
+                        eval.remarks = evaluation.remarks;
+                        eval.companyId = claimresponse.companyId;
+                        _context.evaluations.Add(eval);
+                        await _context.SaveChangesAsync();
+                    }
+                    if (autoId != "")
+                    {
+                        _context.ChangeTracker.Clear();
+                        Evaluation eval = new Evaluation();
+                        string comId = "EVAL" + (int.Parse(autoId) + 1);
+                        eval.evaluationAutoId = evaluation.evaluationAutoId;
+                        eval.evaluationId = comId;
+                        eval.woAutoId = evaluation.woAutoId;
+                        eval.userName = evaluation.userName;
+                        eval.topName = evaluation.topName;
+                        eval.startTime = evaluation.startTime;
+                        eval.endTime = evaluation.endTime;
+                        eval.remarks = evaluation.remarks;
+                        eval.companyId = claimresponse.companyId;
+                        _context.evaluations.Add(eval);
+                        await _context.SaveChangesAsync();
+                    }
+
+                    return Ok();
+                }
+                return Unauthorized();
             }
-
-            if (autoId == "")
+            catch (Exception ex)
             {
-                _context.ChangeTracker.Clear();
-                Evaluation eval = new Evaluation();
-                string comId = "EVAL1";
-                eval.evaluationAutoId = evaluation.evaluationAutoId;
-                eval.evaluationId = comId;
-                eval.woAutoId = evaluation.woAutoId;
-                eval.userName = evaluation.userName;
-                eval.topName =evaluation.topName;
-                eval.startTime = evaluation.startTime;
-                eval.endTime=evaluation.endTime;
-                eval.remarks = evaluation.remarks;
-                eval.companyId = evaluation.companyId;
-                _context.evaluations.Add(eval);
-                await _context.SaveChangesAsync();
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
-            if (autoId != "")
-            {
-                _context.ChangeTracker.Clear();
-                Evaluation eval = new Evaluation();
-                string comId = "EVAL" + (int.Parse(autoId) + 1);
-                eval.evaluationAutoId = evaluation.evaluationAutoId;
-                eval.evaluationId = comId;
-                eval.woAutoId = evaluation.woAutoId;
-                eval.userName = evaluation.userName;
-                eval.topName = evaluation.topName;
-                eval.startTime = evaluation.startTime;
-                eval.endTime = evaluation.endTime;
-                eval.remarks = evaluation.remarks;
-                eval.companyId = evaluation.companyId;
-                _context.evaluations.Add(eval);
-                await _context.SaveChangesAsync();
-            }
-
-            return Ok();
 
         }
 
@@ -130,16 +196,31 @@ namespace MaintenanceWebApi.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteEvaluation(int id)
         {
-            var evaluation = await _context.evaluations.FindAsync(id);
-            if (evaluation == null)
+            try
             {
-                return NotFound();
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    var evaluation = await _context.evaluations.FindAsync(id);
+                    if (evaluation == null)
+                    {
+                        return NotFound();
+                    }
+
+                    _context.evaluations.Remove(evaluation);
+                    await _context.SaveChangesAsync();
+
+                    return NoContent();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
-            _context.evaluations.Remove(evaluation);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
         }
 
         private bool EvaluationExists(int id)
