+++ MaintenanceWebApi/Controllers/MethodStepsController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MaintenanceWebApi.Models;
+using JwtAuthenticationManager;
+using JwtAuthenticationManager.Models;
 
 namespace MaintenanceWebApi.Controllers
 {
@@ -14,62 +16,108 @@ namespace MaintenanceWebApi.Controllers
     public class MethodStepsController : ControllerBase
     {
         private readonly MaintenanceDbContext _context;
+        private readonly JwtTokenHandler _JwtTokenHandler;
+        private readonly ILogger<MethodStepsController> _logger;
 
-        public MethodStepsController(MaintenanceDbContext context)
+        public MethodStepsController(MaintenanceDbContext context, JwtTokenHandler jwtTokenHandler, ILogger<MethodStepsController> logger)
         {
             _context = context;
+            _JwtTokenHandler = jwtTokenHandler;
+            _logger = logger;
         }
 
         // GET: api/MethodSteps
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<MethodSteps>>> GetmethodSteps(string companyId)
+        public async Task<ActionResult<IEnumerable<MethodSteps>>> GetmethodSteps()
         {
-            return await _context.methodSteps.Where(x=>x.companyId==companyId).ToListAsync();
+            try
+            {
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    return await _context.methodSteps.Where(x => x.companyId == claimresponse.companyId).ToListAsync();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // GET: api/MethodSteps/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<MethodSteps>> GetMethodSteps(int id,string companyId)
+        public async Task<ActionResult<MethodSteps>> GetMethodSteps(int id)
         {
-            var methodSteps = await _context.methodSteps.Where(x=>x.msAutoId==id && x.companyId==companyId).ToListAsync();
-
-            if (methodSteps == null)
+            try
             {
-                return NotFound();
-            }
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
 
-            return Ok(methodSteps);
+                    var methodSteps = await _context.methodSteps.Where(x => x.msAutoId == id && x.companyId == claimresponse.companyId).ToListAsync();
+
+                    if (methodSteps == null)
+                    {
+                        return NotFound();
+                    }
+
+                    return Ok(methodSteps);
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // PUT: api/MethodSteps/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut]
-        public async Task<IActionResult> PutMethodSteps(int id, string companyId, MethodSteps methodSteps)
+        public async Task<IActionResult> PutMethodSteps(int id, MethodSteps methodSteps)
         {
-            if (methodSteps.msAutoId==id && methodSteps.companyId==companyId)
-            {
-                _context.Entry(methodSteps).State = EntityState.Modified;
-            }
-
-            
-
             try
             {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!MethodStepsExists(id))
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
 
-            return NoContent();
+                    if (methodSteps.msAutoId == id && methodSteps.companyId == claimresponse.companyId)
+                    {
+                        _context.Entry(methodSteps).State = EntityState.Modified;
+                    }
+
+                    try
+                    {
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!MethodStepsExists(id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+
+                    return NoContent();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // POST: api/MethodSteps
@@ -77,26 +125,55 @@ namespace MaintenanceWebApi.Controllers
         [HttpPost]
         public async Task<ActionResult<MethodSteps>> PostMethodSteps(MethodSteps methodSteps)
         {
-            _context.methodSteps.Add(methodSteps);
-            await _context.SaveChangesAsync();
+            try
+            {
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
 
-            return Ok();
+                    _context.methodSteps.Add(methodSteps);
+                    await _context.SaveChangesAsync();
+
+                    return Ok();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // DELETE: api/MethodSteps/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteMethodSteps(int id)
         {
-            var methodSteps = await _context.methodSteps.FindAsync(id);
-            if (methodSteps == null)
+            try
             {
-                return NotFound();
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    var methodSteps = await _context.methodSteps.FindAsync(id);
+                    if (methodSteps == null)
+                    {
+                        return NotFound();
+                    }
+
+                    _context.methodSteps.Remove(methodSteps);
+                    await _context.SaveChangesAsync();
+
+                    return NoContent();
+                }
+                return Unauthorized();
             }
-
-            _context.methodSteps.Remove(methodSteps);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         private bool MethodStepsExists(int id)
