+++ MaintenanceWebApi/Controllers/ProcedureHealthAndSafetiesController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MaintenanceWebApi.Models;
+using JwtAuthenticationManager;
+using JwtAuthenticationManager.Models;
 
 namespace MaintenanceWebApi.Controllers
 {
@@ -14,62 +16,108 @@ namespace MaintenanceWebApi.Controllers
     public class ProcedureHealthAndSafetiesController : ControllerBase
     {
         private readonly MaintenanceDbContext _context;
+        private readonly JwtTokenHandler _JwtTokenHandler;
+        private readonly ILogger<ProcedureHealthAndSafetiesController> _logger;
 
-        public ProcedureHealthAndSafetiesController(MaintenanceDbContext context)
+        public ProcedureHealthAndSafetiesController(MaintenanceDbContext context, JwtTokenHandler jwtTokenHandler, ILogger<ProcedureHealthAndSafetiesController> logger)
         {
             _context = context;
+            _JwtTokenHandler = jwtTokenHandler;
+            _logger = logger;
         }
 
         // GET: api/ProcedureHealthAndSafeties
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<ProcedureHealthAndSafety>>> GetprocedureHealthAndSafeties(string companyId)
+        public async Task<ActionResult<IEnumerable<ProcedureHealthAndSafety>>> GetprocedureHealthAndSafeties()
         {
-            return await _context.procedureHealthAndSafeties.Where(x=>x.companyId==companyId).ToListAsync();
+            try
+            {
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    return await _context.procedureHealthAndSafeties.Where(x => x.companyId == claimresponse.companyId).ToListAsync();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // GET: api/ProcedureHealthAndSafeties/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<ProcedureHealthAndSafety>> GetProcedureHealthAndSafety(int id,string companyId)
+        public async Task<ActionResult<ProcedureHealthAndSafety>> GetProcedureHealthAndSafety(int id)
         {
-            var procedureHealthAndSafety = await _context.procedureHealthAndSafeties.Where(x=>x.phsAutoId==id && x.companyId==companyId).ToListAsync();
-
-            if (procedureHealthAndSafety == null)
+            try
             {
-                return NotFound();
-            }
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    var procedureHealthAndSafety = await _context.procedureHealthAndSafeties.Where(x => x.phsAutoId == id && x.companyId == claimresponse.companyId).ToListAsync();
 
-            return Ok(procedureHealthAndSafety);
+                    if (procedureHealthAndSafety == null)
+                    {
+                        return NotFound();
+                    }
+
+                    return Ok(procedureHealthAndSafety);
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // PUT: api/ProcedureHealthAndSafeties/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut]
-        public async Task<IActionResult> PutProcedureHealthAndSafety(int id,string companyId, ProcedureHealthAndSafety procedureHealthAndSafety)
+        public async Task<IActionResult> PutProcedureHealthAndSafety(int id,ProcedureHealthAndSafety procedureHealthAndSafety)
         {
-            if (procedureHealthAndSafety.phsAutoId ==id && procedureHealthAndSafety.companyId==companyId)
-            {
-                _context.Entry(procedureHealthAndSafety).State = EntityState.Modified;
-
-            }
-
-
             try
             {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!ProcedureHealthAndSafetyExists(id))
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
+                    if (procedureHealthAndSafety.phsAutoId == id && procedureHealthAndSafety.companyId == claimresponse.companyId)
+                    {
+                        _context.Entry(procedureHealthAndSafety).State = EntityState.Modified;
 
-            return NoContent();
+                    }
+
+
+                    try
+                    {
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!ProcedureHealthAndSafetyExists(id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+
+                    return NoContent();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // POST: api/ProcedureHealthAndSafeties
@@ -77,26 +125,54 @@ namespace MaintenanceWebApi.Controllers
         [HttpPost]
         public async Task<ActionResult<ProcedureHealthAndSafety>> PostProcedureHealthAndSafety(ProcedureHealthAndSafety procedureHealthAndSafety)
         {
-            _context.procedureHealthAndSafeties.Add(procedureHealthAndSafety);
-            await _context.SaveChangesAsync();
+            try
+            {
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    _context.procedureHealthAndSafeties.Add(procedureHealthAndSafety);
+                    await _context.SaveChangesAsync();
 
-            return Ok();
+                    return Ok();
+                }
+                return Unauthorized();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // DELETE: api/ProcedureHealthAndSafeties/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteProcedureHealthAndSafety(int id)
         {
-            var procedureHealthAndSafety = await _context.procedureHealthAndSafeties.FindAsync(id);
-            if (procedureHealthAndSafety == null)
+            try
             {
-                return NotFound();
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    var procedureHealthAndSafety = await _context.procedureHealthAndSafeties.FindAsync(id);
+                    if (procedureHealthAndSafety == null)
+                    {
+                        return NotFound();
+                    }
+
+                    _context.procedureHealthAndSafeties.Remove(procedureHealthAndSafety);
+                    await _context.SaveChangesAsync();
+
+                    return NoContent();
+                }
+                return Unauthorized();
             }
-
-            _context.procedureHealthAndSafeties.Remove(procedureHealthAndSafety);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
         }
 
         private bool ProcedureHealthAndSafetyExists(int id)
